time int

+++ PhysicsEngine/Controller/iController.java
@@ -1,5 +1,5 @@
 package PhysicsEngine.Controller;
 
 public interface iController {
-    public double[] getUV(double[][] state, double time);
+    public double[] getUV(double[][] state, int time);
 }
